requirements.txt
streamlit
datasets
transformers
faiss-cpu
requests
bs4
keybert
nltk
sentence_transformers
rouge_score